+++ pages/projects/index.tsx
@@ -1,17 +1,34 @@
 import Projects from '@/src/components/projects'
-import { unstable_getServerSession } from 'next-auth/next'
-import { authOptions } from '../api/auth/[...nextauth]'
 import SideBar from '@/src/components/side-bar'
+import { authOptions } from '../api/auth/[...nextauth]'
 import useSWR from 'swr'
 import { fetcher } from '@/util/api'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
+import { unstable_getServerSession } from 'next-auth/next'
+import { useState } from 'react'
+import { useBreakpointValue } from '@chakra-ui/react'
 
-export default function Page ({ session }): JSX.Element {
+const smVariant = { navigation: 'drawer', navigationButton: true }
+const mdVariant = { navigation: 'sidebar', navigationButton: false }
+
+export default function Page ({ session }: { session: any }): JSX.Element {
   const { data, error, mutate } = useSWR('/api/projects', fetcher)
   const { data: industries, error: errorIndustries } = useSWR('/api/industries', fetcher)
+  const [isSidebarOpen, setSidebarOpen] = useState(false)
+  const variants = useBreakpointValue({ base: smVariant, md: mdVariant })
+
+  const toggleSidebar = (): void => setSidebarOpen(!isSidebarOpen)
+
   return (
-    <SideBar page='projects'>
-      <Projects projects={data || []} session={session} fetchProjects={mutate} />
+    <SideBar
+      page='projects'
+      variant={variants?.navigation}
+      isOpen={isSidebarOpen}
+      onClose={toggleSidebar}
+      showSidebarButton={variants?.navigationButton}
+      onShowSidebar={toggleSidebar}
+    >
+      <Projects projects={data ?? []} session={session} fetchProjects={mutate} />
     </SideBar>
   )
 }
@@ -19,7 +36,7 @@ export default function Page ({ session }): JSX.Element {
 export async function getServerSideProps (ctx): Promise<any> {
   const session = await unstable_getServerSession(ctx.req, ctx.res, authOptions)
 
-  if (session == null) {
+  if (session?.user == null) {
     return {
       redirect: {
         destination: '/login',
